Signin and Signup are in the same SignCard component
